InfoBubbleを表示する (#15)
* wip

* Fix: InfoWindow positioning is incorrect (confirmed: Google Maps, Mapbox, and Here. Not confirmed yet: ArcGIS)

* Info bubbles follow the dragged markers.

* update: .gitignore file

* wip: Implementing the marker drag gesture for Here SDK

* fix: linter errors

* fix: duplicated markers are added

* fix: KtLint errors

* update README.md

* The mapProvider of MapViewBase is no longer necessary

* Implement toasting logic in AppViewModel

* Create Settings class for defining default values

* rename MapView.kt to MapViewContainer.kt

* Fix: Cannot follow the map camera position for ArcGIS MapView

* Fix: InfoBubble does not follow the clicked marker on ArcGIS MapView

* Fix: onMapClick does not be involved

* Optimize code structure

* Fix: InfoBubble does not follow on ArcGIS Mapview when device is rotated

* Fix: linter errors

* Launch Google Maps app when taps on the "Get Directions" button in InfoBubble

+++ mapconductor-core/src/main/java/com/mapconductor/core/geocell/HexCellRegistry.kt
@@ -17,7 +17,7 @@ class HexCellRegistry<T : IdentifiedPoint>(
     private val allEntries = ConcurrentHashMap<String, String>()
     private var needsRebuild = false
 
-//    fun registerPoints(points: List<T>) {
+    //    fun registerPoints(points: List<T>) {
 //        allCells.clear()
 //        points.forEach {
 //            val coord = geocell.latLngToHex(it.point, zoom)
@@ -154,7 +154,7 @@ class HexCellRegistry<T : IdentifiedPoint>(
 
         val dx = p2.x - p1.x
         val dy = p2.y - p1.y
-        return sqrt(dx * dx + dy * dy)
+        return sqrt(dx * dx + dy * dy).toDouble()
     }
 
     fun findWithinPixelRadius(
